Experiencia de Aprendizaje 2 - Desarrollo de Agente Inteligentes con AI/Clase 2.3/requirements.txt
langgraph
langchain-openai
langchain-mongodb
pymongo
python-dotenv
openai
streamlit
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
langgraph-cli[inmem]
